op-test: Move SOL monitor thread to seperate file
We now can monitor SOL either with timeout or not.
