+++ common/OpTestThread.py
@@ -156,113 +156,3 @@ class OpSSHThreadRandom(threading.Thread):
         log.debug("Thread exiting after run for desired time")
 
 
-class OpSOLMonitorThread1(threading.Thread):
-    '''
-    This thread just monitors the SOL console for any failures when tests are running
-    on other SSH threads
-    '''
-    def __init__(self, threadID, name, execution_time):
-        threading.Thread.__init__(self)
-        self.threadID = threadID
-        self.name = name
-        self.execution_time = execution_time
-        conf = OpTestConfiguration.conf
-        self.system = conf.system()
-        self.system.goto_state(OpSystemState.OS)
-        self.c = self.system.console.get_console()
-
-    def run(self):
-        log.debug("Starting %s" % self.name)
-        self.sol_monitor_thread(self.name, self.execution_time)
-        log.debug("Exiting %s" % self.name)
-
-    def sol_monitor_thread(self, threadName, torture_time):
-        execution_time = time.time() + 60*torture_time,
-        log.debug("Starting %s for new SOL thread" % threadName)
-        while True:
-            try:
-                self.c.expect("\n", timeout=60)
-            except pexpect.TIMEOUT:
-                pass
-            except pexpect.EOF:
-                self.c.state = IPMIConsoleState.DISCONNECTED
-                self.c = self.system.console.get_console()
-            if time.time() > execution_time:
-                break
-        log.debug("Thread exiting after run for desired time")
-
-class OpSOLMonitorThread2(threading.Thread):
-    '''
-    This thread monitors SOL console by running set of debug commands i.e
-    # Monitor for errors
-    cmd_list = ["grep ',[0-4]\]' /sys/firmware/opal/msglog", "dmesg --color=never -T --level=alert,crit,err,warn",
-                "dmesg -T --level=alert,crit,err,warn"]
-    '''
-    def __init__(self, threadID, name, cmd_list, sleep_time, execution_time, ignore_fail=False):
-        threading.Thread.__init__(self)
-        self.threadID = threadID
-        self.name = name
-        self.cmd_list = cmd_list
-        self.sleep_time = sleep_time
-        self.execution_time = execution_time
-        self.ignore_fail = ignore_fail
-        conf = OpTestConfiguration.conf
-        self.system = conf.system()
-        self.system.goto_state(OpSystemState.OS)
-        self.c = self.system.console
-
-    def run(self):
-        log.debug("Starting %s" % self.name)
-        self.sol_monitor_thread(self.name, self.cmd_list, self.sleep_time, self.execution_time, self.ignore_fail)
-        log.debug("Exiting %s" % self.name)
-
-    def sol_monitor_thread(self, threadName, cmd_list, sleep_time, torture_time, ignore_fail):
-        execution_time = time.time() + 60*torture_time,
-        log.debug("Starting %s for new SOL thread %s" % (threadName, cmd_list))
-        while True:
-            for cmd in cmd_list:
-                if ignore_fail:
-                    try:
-                        self.c.run_command(cmd)
-                    except CommandFailed as cf:
-                        pass
-                else:
-                    self.c.run_command(cmd)
-                time.sleep(sleep_time)
-            if time.time() > execution_time:
-                break
-        log.debug("Thread exiting after run for desired time")
-
-
-class OpSolMonitorThread3(threading.Thread):
-    '''
-    This thread just monitors the SOL console for any failures when tests are running
-    on other SSH threads. This thread can be terminated by just calling console_terminate
-    from parent process.
-    '''
-    def __init__(self):
-        threading.Thread.__init__(self)
-        conf = OpTestConfiguration.conf
-        self.system = conf.system()
-        self.system.goto_state(OpSystemState.OS)
-        self.c = self.system.console.get_console()
-        self.c_terminate = False;
-
-
-    def run(self):
-        log.debug("Starting SOL monitoring thread")
-        while True:
-            try:
-                self.c.expect("\n", timeout=60)
-            except pexpect.TIMEOUT:
-                pass
-            except pexpect.EOF:
-                self.c.state = IPMIConsoleState.DISCONNECTED
-                self.c = self.system.console.get_console()
-
-            if self.c_terminate:
-                break
-        log.debug("Terminating SOL monitoring thread")
-
-    def console_terminate(self):
-        self.c_terminate = True
